+++ app/src/main/java/org/kontalk/ui/MessagingNotification.java
@@ -978,10 +978,11 @@ public class MessagingNotification {
                     for (NotificationConversation.ConversationMessage message : content) {
                         Contact contact = Contact.findByUserId(mContext, message.peer);
                         name = contact.getDisplayName();
+                        String uri = contact.getUri() != null ? contact.getUri().toString() : null;
 
                         Person person = new Person.Builder()
                             .setName(name)
-                            .setUri(contact.getUri().toString())
+                            .setUri(uri)
                             .setKey(contact.getJID())
                             .setIcon(IconCompat.createWithBitmap(contact.getAvatarBitmap(mContext, true)))
                             .build();
@@ -1043,9 +1044,18 @@ public class MessagingNotification {
 
                 // group avatar and ticker
                 if (conv.groupJid != null) {
-                    mBuilder.setLargeIcon(MessageUtils
-                        .drawableToBitmap(ContextCompat
-                            .getDrawable(mContext, R.drawable.ic_default_group)));
+                    // default group avatar
+                    Bitmap defaultGroupAvatar = MessageUtils.drawableToBitmap(ContextCompat
+                            .getDrawable(mContext, R.drawable.ic_default_group));
+
+                    if (Contact.isRoundedAvatars()) {
+                        Bitmap roundedAvatar = MediaStorage.createRoundBitmap(defaultGroupAvatar);
+                        if (roundedAvatar != defaultGroupAvatar)
+                            defaultGroupAvatar.recycle();
+                        defaultGroupAvatar = roundedAvatar;
+                    }
+
+                    mBuilder.setLargeIcon(defaultGroupAvatar);
                 }
 
                 SpannableStringBuilder buf = new SpannableStringBuilder();